only show favorite icon on activity grid
